requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
python-dotenv==1.0.1
jinja2==3.1.4
Pillow>=10.0.0
google-genai>=0.4.0
